+++ CHANGELOG.md
@@ -36,7 +36,7 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 - updated Polish translation (Thorsten)
 - updated French translation (Thorsten)
 - updated Japanese translation (Kamo Tomoyuki)
-- updated Brazilian Portuguese translation (Thorsten)
+- updated Portuguese and Brazilian Portuguese translation (Thorsten)
 - updated Turkish translation (Thorsten)
 - updated Dutch translation (Thorsten)
 - updated Norwegian (Bokmål) translation (Thorsten)